+++ tests/electron/modules/calendar_spec.js
@@ -26,12 +26,14 @@ describe("Calendar module", () => {
 		});
 
 		it("has css class today", async () => {
-			await helpers.startApplication("tests/configs/modules/calendar/custom.js", "01 Jan 2030 12:30:00 GMT");
+			helpers.mockSystemDate("01 Jan 2030 12:30:00 GMT");
+			await helpers.startApplication("tests/configs/modules/calendar/custom.js");
 			await doTest(".today");
 		});
 
 		it("has css class tomorrow", async () => {
-			await helpers.startApplication("tests/configs/modules/calendar/custom.js", "31 Dec 2029 12:30:00 GMT");
+			helpers.mockSystemDate("31 Dec 2029 12:30:00 GMT");
+			await helpers.startApplication("tests/configs/modules/calendar/custom.js");
 			await doTest(".tomorrow");
 		});