+++ CHANGELOG.md
@@ -191,6 +191,7 @@ Special thanks to @khassel, @rejas and @sdetweil for taking over most (if not al
   - Reworked how weather-providers handle units (#2849)
   - Use unix() method for parsing times, fix suntimes on the way (#2950)
   - Refactor conversion functions into utils class (#2958)
+- Use async/await for startup of the application instead of callbacks
 - The `cors`-method in `server.js` now supports sending and receiving HTTP headers
 - Replace `&hellip;` by `…`
 - Cleanup compliments module